+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-+ Ensure both delays are written for the two SAR dates in the GUNW
++ Add assert statement to raise error if the delay cube for each SAR date in a GUNW IFG is not written 
 
 ## [0.4.1]